fix: returns 400 bad request on case creation with name already in use
